feat(generate_index.sh): Added function to generate added.tmp
File "added.tmp" includes newly added songs which are not uploaded
and that are found in the last execution of app to sync music files.
I added the function which compares list of songs already uploaded
(Index) and list of songs in given directory (index.tmp). As a result,
songs which are unique to "index.tmp" are added in "added.tmp".

usage: added.tmp would be used to get the list of songs that are still
       needed to be added.