chore(setup.sh): Updated setup.sh with pipenv
Because travis-ci builds failed with complete paths. Thus, along with
adding pipenv, I am changing some complete paths to relative path,
keeping in mind that setup.sh is needed to be executed from
application directory. With this change, travis builds won't fail with
error of "file/directory not found" for path of files changed in this
commit.