feat(*): implemented .env for handling dropbox key
Adding the usage of .env file. Since the project is using pipenv, which
automatically loads this file. So, implementing this makes storing and
retrieving of configurtion variable such as dropbox auth2 token in
environment becomes really handy than the previous approach of using a
keys.json. With this commit, we no longer need keys.json file and
functions that implemented its functionality.